+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 BBC iPlayer Change Log
 ======================
 
+1.3 - unreleased
+----------------
+
+  * Fix some integration issues with radio shows.
+  * Integrate playback position.
+  * Integrate volume controls.
+
 1.2 - September 1, 2017
 ---------------------